Overlay List Screenshot Functionality

+++ app/src/main/java/com/overlayscreendesigntest/data/ResultGroup.kt
@@ -3,7 +3,7 @@ package com.overlayscreendesigntest.data
 data class ResultGroup(
     val average_score: Double,
     val detected_item: DetectedItem,
-    val max_score: Int,
+    val max_score: Double,
     val rank_score: Double,
     val similar_products: List<SimilarProduct>
 )
